requirements.txt
chinese_calendar==1.6.0
openpyxl==3.0.7
pandas==1.2.4
pdfminer3k==1.3.4
pytest==6.2.3
python_dateutil==2.8.2
python_docx==0.8.11
PyYAML==6.0
requests==2.25.1
urllib3==1.26.5
deepdiff==5.7.0
Deprecated==1.2.13
PyYAML==6.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
chardet==4.0.0
chinese-calendar==1.6.0
deepdiff==5.7.0
et-xmlfile==1.1.0
idna==2.10
importlib-metadata==4.11.4
iniconfig==1.1.1
lxml==4.9.0
numpy==1.21.6
openpyxl==3.0.7
ordered-set==4.0.2
packaging==21.3
pandas==1.2.4
pdfminer3k==1.3.4
pip==22.1.2
pluggy==0.13.1
ply==3.11
py==1.11.0
pyparsing==3.0.9
pytest==6.2.3
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
redis==4.3.4
requests==2.25.1
setuptools==47.1.0
six==1.16.0
toml==0.10.2
typing-extensions==4.2.0
urllib3==1.26.5
wrapt==1.14.1
zipp==3.8.0


